requirements.txt
click==8.1.7
hydra-core==1.3.2
matplotlib==3.8.2
omegaconf==2.3.0
pytorch_lightning==2.1.3
torch==2.1.2
wandb == 0.16.2